+++ features/mbedtls/targets/TARGET_NUVOTON/TARGET_M480/des/des_alt.c
@@ -14,9 +14,6 @@
  * limitations under the License.
  */
 
-/* Compatible with mbed OS 2 which doesn't support mbedtls */
-#if MBED_CONF_RTOS_PRESENT
-
 #if !defined(MBEDTLS_CONFIG_FILE)
 #include "mbedtls/config.h"
 #else
@@ -31,7 +28,7 @@
 #include "des_alt.h"
 #include "crypto-misc.h"
 #include "nu_bitutil.h"
-#include "toolchain.h"
+#include "mbed_toolchain.h"
 
 // Must be a multiple of 64-bit block size
 #define MAXSIZE_DMABUF  (8 * 5)
@@ -411,6 +408,3 @@ static int mbedtls_des_docrypt(uint16_t keyopt, uint8_t key[3][MBEDTLS_DES_KEY_S
 
 #endif /* MBEDTLS_DES_ALT */
 #endif /* MBEDTLS_DES_C */
-
-#endif /* MBED_CONF_RTOS_PRESENT */
-
