+++ features/mbedtls/targets/TARGET_NUVOTON/TARGET_M480/des/des_alt.h
@@ -23,6 +23,7 @@
 #include MBEDTLS_CONFIG_FILE
 #endif
 
+#if defined(MBEDTLS_DES_C)
 #if defined(MBEDTLS_DES_ALT)
 
 #include <stddef.h>
@@ -276,5 +277,6 @@ void mbedtls_des_setkey( uint32_t SK[32],
 #endif
 
 #endif /* MBEDTLS_DES_ALT */
+#endif /* MBEDTLS_DES_C */
 
 #endif /* des_alt.h */
