+++ features/mbedtls/targets/TARGET_NUVOTON/TARGET_M480/sha/sha256_alt_sw.c
@@ -24,9 +24,6 @@
  *  http://csrc.nist.gov/publications/fips/fips180-2/fips180-2.pdf
  */
 
-/* Compatible with mbed OS 2 which doesn't support mbedtls */
-#if MBED_CONF_RTOS_PRESENT
-
 #if !defined(MBEDTLS_CONFIG_FILE)
 #include "mbedtls/config.h"
 #else
@@ -318,7 +315,4 @@ void mbedtls_sha256_sw_finish( mbedtls_sha256_sw_context *ctx, unsigned char out
 }
 
 #endif /* MBEDTLS_SHA256_ALT */
-
 #endif /* MBEDTLS_SHA256_C */
-
-#endif /* MBED_CONF_RTOS_PRESENT */
